Formatted code and renamed map page
