+++ final_project/lib/pages/main_scaffold.dart
@@ -6,7 +6,7 @@ import "package:provider/provider.dart";
 
 import "../widgets/authentication.dart";
 import "/pages/body/about.dart";
-import "/pages/body/home.dart";
+import "body/map.dart";
 import "/pages/report.dart";
 
 class MainScaffold extends StatefulWidget {
@@ -60,7 +60,8 @@ class _MainScaffoldState extends State<MainScaffold> {
       bottomNavigationBar: BottomNavigationBar(
         backgroundColor: const Color.fromARGB(255, 139, 69, 19),
         items: const [
-          BottomNavigationBarItem(icon: Icon(Icons.leaderboard), label: "Leaderboard"),
+          BottomNavigationBarItem(
+              icon: Icon(Icons.leaderboard), label: "Leaderboard"),
           BottomNavigationBarItem(icon: Icon(Icons.home), label: "Home"),
           BottomNavigationBarItem(icon: Icon(Icons.info), label: "About"),
         ],
@@ -78,13 +79,10 @@ class _MainScaffoldState extends State<MainScaffold> {
       // TODO: see if creating these page bodies every time matters
       var (pageBody, pageTitle) = switch (index) {
         0 => (
-            Leaderboard(), 
+            Leaderboard(),
             "Leaderboard"
           ), // TODO: replace this null when you create the Leaderboard body widget
-        1 => (
-            Home(),
-            "Hendrix Arboretum"
-          ), // TODO: replace this null when you create the main body widget
+        1 => (Map(), "Hendrix Arboretum"),
         2 => (About(), "About"),
         _ => (null, null)
       };
